Refactor Friends widget into seperate component on Ascent Modal
